Add functions to fix a bug in the algorithm frontend implementation
Before 2020-04-17, when a respondent declared a major gravity
symptom (i.e. "breathlessness" and/or "feeding_day" being "true"),
those symptomes were both wrongly stored as "false" because the
frontend application discarded the declared value.

This manual fix consists in preserving the orientation message in
those case: this message should always stay "SAMU", if when both
"breathlessness" and/or "feeding_day" are "false".

Also, this commit normalize the orientation messages themselves
by using the shorter 2020-04-29 notation, even for entries stored
before that date -- the reason for this inconsistency is that the
institution we share data with want a normalized version for the
whole dataset.